+++ kobay.com/src/main/webapp/WEB-INF/config/springmvc/dispatcher-servlet.xml
@@ -12,6 +12,13 @@
         <context:exclude-filter type="annotation" expression="org.springframework.stereotype.Service"/>
         <context:exclude-filter type="annotation" expression="org.springframework.stereotype.Repository"/>
     </context:component-scan>
+    
+    <!-- For Json s  -->
+	<mvc:annotation-driven/>
+	<context:annotation-config/>
+	<bean id="jsonView" class="org.springframework.web.servlet.view.json.MappingJackson2JsonView"/>
+	<bean id="beanNameResolver" class="org.springframework.web.servlet.view.BeanNameViewResolver"/>
+	<!-- For Json e -->
 
     <bean class="org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerAdapter">
         <property name="webBindingInitializer">
